QE-1039: Refactor API Auth token simple
