+++ application/libraries/Api/Authentication/AuthenticationTokenSimple.php
@@ -8,12 +8,23 @@ use LSUserIdentity;
 use PluginEvent;
 use CDbCriteria;
 use Session;
-use User;
 use Yii;
 
 class AuthenticationTokenSimple implements AuthenticationInterface
 {
     const ERROR_INVALID_SESSION_KEY = 'INVALID_SESSION_KEY';
+
+    protected SessionUtil $sessionUtil;
+
+    /**
+     * Constructor
+     *
+     * @param SessionUtil $sessionUtil
+     */
+    public function __construct(SessionUtil $sessionUtil)
+    {
+        $this->sessionUtil = $sessionUtil;
+    }
 
     /**
      * Login with username and password
@@ -53,7 +64,7 @@ class AuthenticationTokenSimple implements AuthenticationInterface
      */
     public function createSession($username)
     {
-        $this->jumpStartSession($username);
+        $this->sessionUtil->jumpStartSession($username);
         $sessionKey = (string) Yii::app()->securityManager
             ->generateRandomString(32);
         $session = new Session();
@@ -62,49 +73,6 @@ class AuthenticationTokenSimple implements AuthenticationInterface
         $session->data = $username;
         $session->save();
         return $session;
-    }
-
-    /**
-     * Fills the session with necessary user info on the fly
-     *
-     * @param string $username The username
-     * @return bool
-     */
-    public function jumpStartSession($username)
-    {
-        $oUser = User::model()->findByAttributes(
-            array('users_name' => $username)
-        );
-
-        if (!$oUser) {
-            return false;
-        }
-
-        $aUserData = $oUser->attributes;
-
-        /** @var \LSYii_Application */
-        $app = Yii::app();
-
-        $session = array(
-            'loginID' => intval($aUserData['uid']),
-            'user' => $aUserData['users_name'],
-            'full_name' => $aUserData['full_name'],
-            'htmleditormode' => $aUserData['htmleditormode'],
-            'templateeditormode' => $aUserData['templateeditormode'],
-            'questionselectormode' => $aUserData['questionselectormode'],
-            // When using the REST API, data is transferred using the format
-            // YYYY-MM-DD since the browser handles formatting for display.
-            // This format is defined as '6' in
-            // insurveytranslator_helper.php / getDateFormatData()
-            'dateformat' => 6,
-            'adminlang' => 'en'
-        );
-        foreach ($session as $k => $v) {
-            $app->session[$k] = $v;
-        }
-        $app->user->setId($aUserData['uid']);
-
-        return true;
     }
 
     /**
@@ -126,7 +94,7 @@ class AuthenticationTokenSimple implements AuthenticationInterface
         if (is_null($oResult)) {
             return false;
         } else {
-            $this->jumpStartSession($oResult->data);
+            $this->sessionUtil->jumpStartSession($oResult->data);
             return true;
         }
     }
